random API 추가
