+++ route/view_random.py
@@ -1,0 +1,9 @@
+from .tool.func import *
+
+from .go_api_w_random import api_w_random
+
+def view_random(db_set):
+    with get_db_connect() as conn:
+        data = json.loads(api_w_random(db_set).data)["data"]
+        
+        return redirect('/w/' + url_pas(data))
