+++ route_go/main.go
@@ -14,8 +14,8 @@ func main() {
 		route.Api_w_raw(call_arg[1:])
 	} else if call_arg[0] == "api_func_sha224" {
 		route.Api_func_sha224(call_arg[1:])
-	} else if call_arg[0] == "view_random" {
-		route.View_random(call_arg[1:])
+	} else if call_arg[0] == "api_w_random" {
+		route.Api_w_random(call_arg[1:])
 	} else if call_arg[0] == "api_search" {
 		route.Api_search(call_arg[1:])
 	}
